HV-1667 Create an interface for ConstraintValidatorManager
